+++ src/mathlib/numbers.py
@@ -1,37 +1,41 @@
-import math
 from typing import Iterator
 
+from mathlib import _seven
 
-def lcm(a: int, b: int) -> int:
+
+def gcd(*integers: int) -> int:
     """
-    Find the least common multiple of a and b.
+    Find the greatest common divisor of the arguments.
+
+    The greatest common divisor of a and 0 is always a, as this
+    coincides with gcd to be the greatest lower bound in the
+    lattice of divisibility.
+    """
+    return _seven.gcd(*integers)
+
+
+def lcm(*integers: int) -> int:
+    """
+    Find the least common multiple of the arguments.
 
     The least common multiple of a and 0 is always 0, as this
     coincides with lcm to be the least upper bound in the
     lattice of divisibility.
     """
-    return a * b // math.gcd(a, b)
+    return _seven.lcm(*integers)
 
 
 def modular_inverse(n: int, mod: int) -> int:
     """
     Find the modular inverse of n modulo mod.
 
-    The algorithm is based on using the Extended Euclid Algorithm,
-    to solve the diophantine equation n*x + mod*y = 1.
+    In python 3.6 and 3.7, the algorithm is based on using
+    the Extended Euclid Algorithm, to solve the diophantine
+    equation n*x + mod*y = 1.
+
+    In later versions this is just a wrapper over the pow function.
     """
-    original_modulo = mod
-    x = 1
-    y = 0
-
-    while n > 1:
-        x, y = y, x - (n // mod) * y
-        n, mod = mod, n % mod
-
-    if x < 0:
-        x += original_modulo
-
-    return x
+    return _seven.modular_inverse(n, mod)
 
 
 def fibonacci(n: int, a: int = 1, b: int = 1) -> int: